better field aliasing, ts helpers
