+++ example/lib/graphql/_home_fragment.dart
@@ -1,0 +1,15 @@
+import './schema.dart';
+
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+import 'package:meta/meta.dart';
+import 'package:json_annotation/json_annotation.dart';
+
+///
+mixin Home on Human {
+  static final String typeName = "Human";
+
+  @JsonKey(required: false, disallowNullValue: false)
+  String get home => fields.homePlanet;
+  set home(String home) => fields.homePlanet = home;
+}
