+++ example/lib/graphql/_home_fragment.gql
@@ -1,0 +1,3 @@
+fragment home on Human {
+  home: homePlanet
+}
