+++ src/helpers/multiline-comment.ts
